Move restartGame into main.js
