+++ levels.js
@@ -23,151 +23,4 @@ export const GAME_LEVELS = [`
 ..............................#####..................#######....................
 ................................................................................
 ................................................................................
-`,`                                                                     
-................................................................................
-................................................................................
-....###############################.............................................
-...##.............................##########################################....
-...#.......................................................................##...
-...#....o...................................................................#...
-...#................................................=.......................#...
-...#.o........################...................o..o...........|........o..#...
-...#.........................#..............................................#...
-...#....o....................##########.....###################....##########...
-...#..................................#+++++#.................#....#............
-...###############....oo......=o.o.o..#######.###############.#....#............
-.....#...............o..o.............#.......#......#........#....#............
-.....#....................#############..######.####.#.########....########.....
-.....#.............########..............#...........#.#..................#.....
-.....#..........####......####...#####################.#..................#.....
-.....#........###............###.......................########....########.....
-.....#.......##................#########################......#....#............
-.....#.......#................................................#....#............
-.....###......................................................#....#............
-.......#...............o...........................................#............
-.......#...............................................o...........#............
-.......#########......###.....############.........................##...........
-.............#..................#........#####....#######.o.........########....
-.............#++++++++++++++++++#............#....#.....#..................#....
-.............#++++++++++++++++++#..........###....###...####.o.............#....
-.............####################..........#........#......#.....|.........#....
-...........................................#++++++++#......####............#....
-...........................................#++++++++#.........#........@...#....
-...........................................#++++++++#.........##############....
-...........................................##########...........................
-................................................................................
-`,`
-......................................#++#........................#######....................................#+#..
-......................................#++#.....................####.....####.................................#+#..
-......................................#++##########...........##...........##................................#+#..
-......................................##++++++++++##.........##.............##...............................#+#..
-.......................................##########++#.........#....................................o...o...o..#+#..
-................................................##+#.........#.....o...o....................................##+#..
-.................................................#+#.........#................................###############++#..
-.................................................#v#.........#.....#...#........................++++++++++++++##..
-.............................................................##..|...|...|..##............#####################...
-..............................................................##+++++++++++##............v........................
-...............................................................####+++++####......................................
-...............................................#.....#............#######........###.........###..................
-...............................................#.....#...........................#.#.........#.#..................
-...............................................#.....#.............................#.........#....................
-...............................................#.....#.............................##........#....................
-...............................................##....#.............................#.........#....................
-...............................................#.....#......o..o.....#...#.........#.........#....................
-...............#######........###...###........#.....#...............#...#.........#.........#....................
-..............##.....##.........#...#..........#.....#.....######....#...#...#########.......#....................
-.............##.......##........#.o.#..........#....##...............#...#...#...............#....................
-.....@.......#.........#........#...#..........#.....#...............#...#...#...............#....................
-....###......#.........#........#...#..........#.....#...............#...#####...######......#....................
-....#.#......#.........#.......##.o.##.........#.....#...............#.....o.....#.#.........#....................
-++++#.#++++++#.........#++++++##.....##++++++++##....#++++++++++.....#.....=.....#.#.........#....................
-++++#.#++++++#.........#+++++##.......##########.....#+++++++##+.....#############.##..o.o..##....................
-++++#.#++++++#.........#+++++#....o.................##++++++##.+....................##.....##.....................
-++++#.#++++++#.........#+++++#.....................##++++++##..+.....................#######......................
-++++#.#++++++#.........#+++++##.......##############++++++##...+..................................................
-++++#.#++++++#.........#++++++#########++++++++++++++++++##....+..................................................
-++++#.#++++++#.........#++++++++++++++++++++++++++++++++##.....+..................................................
-`,`
-..............................................................................................................
-..............................................................................................................
-..............................................................................................................
-..............................................................................................................
-..............................................................................................................
-........................................o.....................................................................
-..............................................................................................................
-........................................#.....................................................................
-........................................#.....................................................................
-........................................#.....................................................................
-........................................#.....................................................................
-.......................................###....................................................................
-.......................................#.#.................+++........+++..###................................
-.......................................#.#.................+#+........+#+.....................................
-.....................................###.###................#..........#......................................
-......................................#...#.................#...oooo...#.......###............................
-......................................#...#.................#..........#......#+++#...........................
-......................................#...#.................############.......###............................
-.....................................##...##......#...#......#................................................
-......................................#...#########...########..............#.#...............................
-......................................#...#...........#....................#+++#..............................
-......................................#...#...........#.....................###...............................
-.....................................##...##..........#.......................................................
-......................................#...#=.=.=.=....#............###........................................
-......................................#...#...........#...........#+++#.......................................
-......................................#...#....=.=.=.=#.....o......###.......###..............................
-.....................................##...##..........#.....................#+++#.............................
-..............................o...o...#...#...........#.....#................##v........###...................
-......................................#...#...........#..............#.................#+++#..................
-.............................###.###.###.###.....o.o..#++++++++++++++#...................v#...................
-.............................#.###.#.#.###.#..........#++++++++++++++#........................................
-.............................#.............#...#######################........................................
-.............................##...........##.........................................###......................
-..###.........................#.....#.....#.........................................#+++#................###..
-..#.#.........................#....###....#..........................................###.................#.#..
-..#...........................#....###....#######........................#####.............................#..
-..#...........................#...........#..............................#...#.............................#..
-..#...........................##..........#..............................#.#.#.............................#..
-..#.......................................#.......|####|....|####|.....###.###.............................#..
-..#................###.............o.o....#..............................#.........###.....................#..
-..#...............#####.......##..........#.............................###.......#+++#..........#.........#..
-..#...............o###o.......#....###....#.............................#.#........###..........###........#..
-..#................###........#############..#.oo.#....#.oo.#....#.oo..##.##....................###........#..
-..#......@..........#.........#...........#++#....#++++#....#++++#....##...##....................#.........#..
-..#############################...........#############################.....################################..
-..............................................................................................................
-..............................................................................................................
-`,`
-..................................................................................................###.#.......
-......................................................................................................#.......
-..................................................................................................#####.......
-..................................................................................................#...........
-..................................................................................................#.###.......
-..........................o.......................................................................#.#.#.......
-.............................................................................................o.o.o###.#.......
-...................###................................................................................#.......
-.......+..o..+................................................#####.#####.#####.#####.#####.#####.#####.......
-.......#.....#................................................#...#.#...#.#...#.#...#.#...#.#...#.#...........
-.......#=.o..#............#...................................###.#.###.#.###.#.###.#.###.#.###.#.#####.......
-.......#.....#..................................................#.#...#.#...#.#...#.#...#.#...#.#.....#.......
-.......+..o..+............o..................................####.#####.#####.#####.#####.#####.#######.......
-..............................................................................................................
-..........o..............###..............................##..................................................
-..............................................................................................................
-..............................................................................................................
-......................................................##......................................................
-...................###.........###............................................................................
-..............................................................................................................
-..........................o.....................................................#......#......................
-..........................................................##.....##...........................................
-.............###.........###.........###.................................#..................#.................
-..............................................................................................................
-.................................................................||...........................................
-..###########.................................................................................................
-..#.........#.o.#########.o.#########.o.##................................................#...................
-..#.........#...#.......#...#.......#...#.................||..................#.....#.........................
-..#..@......#####...o...#####...o...#####.....................................................................
-..#######.....................................#####.......##.....##.....###...................................
-........#=..................=................=#...#.....................###...................................
-........#######################################...#+++++++++++++++++++++###+++++++++++++++++++++++++++++++++++
-..................................................############################################################
-..............................................................................................................
 `];
