+++ main.js
@@ -2,7 +2,9 @@
 import { GAME_LEVELS } from './levels.js';
 import { DOMDisplay } from './dom.js';
 import { runGame } from './gameLogic.js';
+import { restartGame } from './gameLogic.js';
+import { dom } from './domElements.js';
 
 runGame(GAME_LEVELS, DOMDisplay);
 
-restartButton.addEventListener('click', restartGame);
+dom.restartButton.addEventListener('click', restartGame);
